requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
ffmpeg-python==0.2.0
frozenlist==1.8.0
future==1.0.0
idna==3.11
linkify-it-py==2.0.3
markdown-it-py==4.0.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
multidict==6.7.1
nodejs-wheel-binaries==24.13.1
platformdirs==4.9.2
propcache==0.4.1
pygame==2.6.1
Pygments==2.19.2
pytubefix==10.3.6
requests==2.32.5
typing_extensions==4.15.0
uc-micro-py==1.0.3
urllib3==2.6.3
yarl==1.22.0
ytmusicapi==1.11.5
